Hide the booking-page menu when the public site is off, and split the tests
Two fixes from review:

- The site menu was showing on booking pages even when the public site
  feature is turned off. In that state Home and Listings both bounce a
  visitor to the admin login, so the menu was a set of dead links. Now the
  menu is built only when the public site is on (as well as not in an
  iframe), matching every other public-site surface.

- Moved the new site-menu and cart-conflict template tests into their own
  focused files under ticket-page/, keeping the main test file under the
  ~400-line target.

Also recorded a follow-up in TODO.md for anchoring the menu to the booking
page's own listing/group, left out here for its cold-start cost.
